Fix 8-bit buffer

+++ modbus_buf8.h
@@ -4,12 +4,12 @@
 #include <stdint.h>
 
 static inline uint_least8_t modbus_get_buf_byte(const void* buf_in, uint16_t i) {
-    const uint16_t* buf = (const uint16_t*) buf_in;
+    const uint8_t* buf = (const uint8_t*) buf_in;
     return buf[i];
 }
 
 static inline void modbus_set_buf_byte(void* buf_io, uint16_t i, uint_least8_t val) {
-    uint16_t* buf = (uint16_t*) buf_io;
+    uint8_t* buf = (uint8_t*) buf_io;
     buf[i] = val;
 }
